Mapping between http status and gRPC codes (#43)
Mapping between http status and gRPC codes

This is a package with a function to map HTTP status code to a
corresponding gRPC codes. Please note that there isn't a 1 to 1 mapping
so in some cases more than one HTTP code maps to the same gRPC codes and
for some of the HTTP code we don't have a good map in gRPC code
therefore we map them to the UNKNOW code.